[bt#13660] Refactoring tests for inbound PO to make them easier to be extended
